cleanup of old files & consolidated predicates
