+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@mlhaufe/brevity",
-    "version": "0.11.1",
+    "version": "0.12.0",
     "description": "A library for Feature-Oriented Programming",
     "keywords": [
         "Object Algebra",
